+++ src/test/java/UseCase/ReviewManagerTest.java
@@ -20,9 +20,11 @@ public class ReviewManagerTest {
 
     @BeforeClass
     public static void setUp() {
-        WriteUser writeUser = new WriteUser(str1 + "/src/test/res/NormalUser", str1 + "/src/test/res/AdminUser");
+        WriteUser writeUser = new WriteUser(str1 + "/src/test/res/NormalUser",
+                str1 + "/src/test/res/AdminUser");
         WriteReview writeReview = new WriteReview(str1 + "/src/test/res/Review");
-        WriteMovie writeMovie = new WriteMovie(str1 + "/src/test/res/Moviedata/", str1 + "/src/test/res/");
+        WriteMovie writeMovie = new WriteMovie(str1 + "/src/test/res/Moviedata/",
+                str1 + "/src/test/res/");
         InstanceMain.setWriteFileClass(writeUser, writeMovie, writeReview);
         rm = InstanceMain.getReviewManager();
     }
@@ -85,22 +87,6 @@ public class ReviewManagerTest {
         assertEquals(7, rm.getCurrMaxRevId());
     }
 
-    @Test
-    public void writeNewReview() {
-        assertTrue(rm.writeNewReview("ReviewManager2", "lslfj", "content to be deleted", 0));
-        assertEquals(8, rm.getCurrMaxRevId());
-        Object[] expect = {"ReviewManager2", "lslfj", "content to be deleted", 0, 8};
-        Object[] actual = rm.getRevInfoById(8);
-        for (int i = 0; i <= 4; i++){
-            assertEquals(expect[i], actual[i]);
-        }
-        assertEquals(7, rm.getReviewList().size());
-        ReviewSort rsort = new ReviewSort();
-        rsort.sortReviews(rm.getReviewList());
-        assertEquals(7, rm.getReviewList().size());
-        InstanceMain.getWriteReview().deleteReviewFile(8);
-
-    }
 
     @Test
 //    TODO
